requirements.txt
accelerate==0.33.0
bitsandbytes==0.43.3
huggingface-hub==0.24.5
hydra-core==1.3.2
numpy==1.24.1
omegaconf==2.3.0
pandas==1.5.0
peft==0.12.0
pyarrow==15.0.0
python-dotenv==1.0.1
safetensors==0.4.3
scikit-learn==1.3.2
tokenizers==0.20.0
tqdm==4.66.4
transformers==4.45.0
wandb==0.17.5